[fix] backend: handling double in drug pack size

+++ backend/src/main/java/com/healthrx/backend/api/external/DrugPackResponse.java
@@ -14,6 +14,6 @@ public class DrugPackResponse {
     private String gtinCode;
     private String accessibilityCategory;
     private String packType;
-    private Integer packSize;
+    private Double packSize;
     private String packUnit;
 }
